Only remove the markdown-file if it is really empty.

+++ src/main/java/de/ppi/selenium/logevent/report/MarkdownReporter.java
@@ -117,7 +117,8 @@ public class MarkdownReporter implements LogReporter {
                         || EventActions.TEST_FINISHED_WITH_EXCEPTION
                                 .equals(action)) {
                     markdown.close();
-                    if (logOnlyOnError) {
+                    if (logOnlyOnError
+                            && currentReportDir.listFiles().length == 1) {
                         markdownFile.delete();
                     }
                     markdownFile = null;
